feat(stats): Ensure reproducible proto compilation
This commit copies over what we've found works well in kleat.
Notably:
(1) All proto compilation happens in docker so that we are not
dependent on anything installed on the system.
(2) The command make proto regenerates protos, and make checkproto
checks if they are up-to-date.
(3) CI runs make checkproto and fails if they are not up to date.

I did not upgrade any of the proto dependencies here, as confirmed
by the fact that no generated code has changed; I just pinned
everything to the current version. Going forward one can just update
the versions in build/protoc/Dockerfile to bump any dependencies.