refactor(protoc): Use better directories in container
It's customary for mounted filesystems to be in /mnt; let's mount
our files there instead of directly in the root of the filesystem.

Rather than have a top-level staging directory, make this directory
in /tmp. Also create it as part of the genproto script so it's
automatically owned by whichever user is running the script.

Finally the version of kork that we build into the image can go in
/opt/proto rather than at the top-level as /proto.